requirements.txt
pillow>=10.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
